Reverse popup toggle
Previously, Rerouter allowed the user to "Enable Rerouter" from the popup. This had the unintuitive effect of causing Rerouter to be disabled at first install and requiring the user to go in and manually activate Rerouter, after having all ready done so in Safari's settings. This has been reversed, now Rerouter is enabled by default and the toggle asks "Pause Rerouter?".